tests: D2 cross-node witness + extract cascade B3 test into own suite
Address two issues from the final branch review on PR #332:

1. The D2 replication test previously wrote to the desktop core and
   read from the desktop core — a single-node roundtrip, not a P2P
   propagation. Add a test-only /desktop/test-fixture/remote-save-behavior
   bridge endpoint that targets the remote node (fixture.remote_core()),
   then keep the desktop-snapshot read (fixture.desktop_core()). The
   write-on-A -> visible-on-B chain is the real D1/D2 witness.
   The endpoint is guarded by DEFRA_AGENT_TAURI_LIVE=1. A new
   createFixtureHelpers() export in adapter.ts provides the typed
   saveBehaviorConfigOnRemote() call from the test.

2. The B3+C2 cascade it.fails test was in the default live sweep,
   consuming 5-10 min per run with no CI signal (expected failures
   don't alert). Extract it into tests/tauri-driver.live.cascade.test.tsx
   with its own describeLive block. Register it as the "cascade" suite in
   run-live-test.mjs but exclude it from the default sweep array. Add
   test:live:cascade npm script. The remaining interrupt test still runs
   in the default sweep; `within` import removed from interrupt.test.tsx.
