+++ apps/desktop-tauri/package.json
@@ -19,6 +19,7 @@
     "test:live:subagent": "node ./tests/run-live-test.mjs --suite subagent",
     "test:live:replication": "node ./tests/run-live-test.mjs --suite replication",
     "test:live:sad-path": "node ./tests/run-live-test.mjs --suite sad-path",
+    "test:live:cascade": "node ./tests/run-live-test.mjs --suite cascade",
     "smoke:remote-fleet": "node ./scripts/remote-fleet-smoke.mjs",
     "preview": "vite preview",
     "tauri": "tauri"
